+++ test/test_linear_expression.py
@@ -62,8 +62,6 @@ def test_empty_linexpr(m):
 
 
 def test_linexpr_with_wrong_data(m):
-    with pytest.raises(ValueError):
-        LinearExpression(1, m)
 
     with pytest.raises(ValueError):
         LinearExpression(xr.Dataset({"a": [1]}), m)
@@ -101,6 +99,41 @@ def test_linexpr_with_data_without_coords(m):
     data = xr.Dataset({"vars": vars, "coeffs": coeffs})
     expr = LinearExpression(data, m)
     assert_linequal(expr, lhs)
+
+
+def test_linexpr_from_constant_dataarray(m):
+    const = xr.DataArray([1, 2], dims=["dim_0"])
+    expr = LinearExpression(const, m)
+    assert (expr.const == const).all()
+    assert expr.nterm == 0
+
+
+def test_linexpr_from_constant_pandas_series(m):
+    const = pd.Series([1, 2], index=pd.RangeIndex(2, name="dim_0"))
+    expr = LinearExpression(const, m)
+    assert (expr.const == const).all()
+    assert expr.nterm == 0
+
+
+def test_linexpr_from_constant_pandas_dataframe(m):
+    const = pd.DataFrame([[1, 2], [3, 4]], columns=["a", "b"])
+    expr = LinearExpression(const, m)
+    assert (expr.const == const).all()
+    assert expr.nterm == 0
+
+
+def test_linexpr_from_constant_numpy_array(m):
+    const = np.array([1, 2])
+    expr = LinearExpression(const, m)
+    assert (expr.const == const).all()
+    assert expr.nterm == 0
+
+
+def test_linexpr_from_constant_scalar(m):
+    const = 1
+    expr = LinearExpression(const, m)
+    assert (expr.const == const).all()
+    assert expr.nterm == 0
 
 
 def test_repr(m):
@@ -534,15 +567,7 @@ def test_linear_expression_groupby_with_series_false(v):
     groups = pd.Series([1] * 10 + [2] * 10, index=v.indexes["dim_2"])
     groups.name = "dim_2"
     with pytest.raises(ValueError):
-        grouped = expr.groupby(groups).sum()
-
-
-def test_linear_expression_groupby_with_series_false(v):
-    expr = 1 * v
-    groups = pd.Series([1] * 10 + [2] * 10, index=v.indexes["dim_2"])
-    groups.name = "dim_2"
-    with pytest.raises(ValueError):
-        grouped = expr.groupby(groups).sum()
+        expr.groupby(groups).sum()
 
 
 def test_linear_expression_groupby_with_dataframe(v):
@@ -550,11 +575,36 @@ def test_linear_expression_groupby_with_dataframe(v):
     groups = pd.DataFrame(
         {"a": [1] * 10 + [2] * 10, "b": list(range(4)) * 5}, index=v.indexes["dim_2"]
     )
-    grouped = expr.groupby(xr.DataArray(groups)).sum()
+    grouped = expr.groupby(groups).sum()
     index = pd.MultiIndex.from_frame(groups)
     assert "group" in grouped.dims
     assert set(grouped.data.group.values) == set(index.values)
     assert grouped.nterm == 3
+
+
+def test_linear_expression_groupby_with_dataarray(v):
+    expr = 1 * v
+    df = pd.DataFrame(
+        {"a": [1] * 10 + [2] * 10, "b": list(range(4)) * 5}, index=v.indexes["dim_2"]
+    )
+    groups = xr.DataArray(df)
+    grouped = expr.groupby(groups).sum()
+    index = pd.MultiIndex.from_frame(df)
+    assert "group" in grouped.dims
+    assert set(grouped.data.group.values) == set(index.values)
+    assert grouped.nterm == 3
+
+
+def test_linear_expression_groupby_with_dataframe_non_aligned(v):
+    expr = 1 * v
+    groups = pd.DataFrame(
+        {"a": [1] * 10 + [2] * 10, "b": list(range(4)) * 5}, index=v.indexes["dim_2"]
+    )
+    target = expr.groupby(groups).sum()
+
+    groups_non_aligned = groups[::-1]
+    grouped = expr.groupby(groups_non_aligned).sum()
+    assert_linequal(grouped, target)
 
 
 @pytest.mark.parametrize("use_fallback", [True, False])